moved VaultTxRecord  to operations
