+++ src/blox/SimpleVault/hooks/useMetaTxActions.ts
@@ -6,7 +6,7 @@ import { useChain } from '@/hooks/useChain';
 import { useTransactionManager } from '@/hooks/useTransactionManager';
 import { convertBigIntsToStrings } from '@/lib/utils';
 import { NotificationMessage } from '../lib/types';
-import { VaultTxRecord } from '../components/PendingTransaction';
+import { VaultTxRecord } from '../lib/operations';
 import { createVaultMetaTxParams, getStoredMetaTxSettings } from '../SimpleVault.ui';
 import { MetaTransaction } from '../../../particle-core/sdk/typescript/interfaces/lib.index';
